Delegate Packets to Workers (#10)
RakNet stricty requires that RakPeer methods only be used
on a single thread. This requires us to handle threading
the packets ourselves after they have been received.
We've added a NetworkManager service to make this
easy, as well as Handler classes for making it easy
to handle the packets once we've received them.