Receive & Process `ID_LOGIN_REQUEST` Packet (#16)
The `ID_LOGIN_REQUEST` packet is send by the client as the first
step in the process of logging in. It sends the username and the
client version and waits for confirmation that the client can
log in. It replies with `ID_LOGIN_REQUEST_RETURN` and
the client then continues the process.