Register Worlds With Master Server (#46)
In order for the master server to tell the client about world servers
to connect to they need to be tracked. On startup, all worlds will
register themselves with the master server and maintain a
connection in order to communicate. This connection
uses a different port than the client one so that a
firewall can be configured to prevent spoofed
world server packets.