Refactor Packet Union Wrapping & Unwrapping (#52)
Since C# supports reflection metadata, rather than registering C# packets manually,
we can use a new "[PacketID]" attribute. This makes it so the packet is included in
all of the places it previously needed to be added. It also includes it in unit tests
that make sure the struct is configured correctly.