Replace `FOMDataUnion` With Dynamic Packet Buffers (#56)
Since some large packets are about to balloon the FOMDataUnion, this makes it
so that packets are handled directly without the use of a union. This is done by
making use of re-usable buffers to read and write packet data directly into.
This avoids increasing the size of all packets while still allowing for the
very large ones to be passed around as well.